Updated pyrad2pyrad-sr options
